Move internal types to pkg
