+++ pkg/config/config.go
@@ -26,7 +26,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/json-iterator/go"
 )
