+++ pkg/config/configmanager.go
@@ -18,7 +18,7 @@
 package config
 
 import (
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
 )
