+++ pkg/config/configmanager_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 )
 
 func Test_convertClusterHealthCheck(t *testing.T) {
