+++ pkg/router/routeruleimpl.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
 	multimap "github.com/jwangsadinata/go-multimap/slicemultimap"
 	//"github.com/alipay/sofa-mosn/pkg/protocol"
