+++ pkg/router/virtualhostimpl.go
@@ -21,7 +21,7 @@ import (
 	"errors"
 	"regexp"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
 	"github.com/alipay/sofa-mosn/pkg/types"
 	"github.com/markphelps/optional"
