+++ pkg/server/config/proxy/generic.go
@@ -20,7 +20,7 @@ package proxy
 import (
 	"context"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/proxy"
 	"github.com/alipay/sofa-mosn/pkg/types"
 )
