+++ pkg/upstream/cluster/clustermanager.go
@@ -24,7 +24,7 @@ import (
 	"net"
 	"sync"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
 	proto "github.com/alipay/sofa-mosn/pkg/protocol"
 	"github.com/alipay/sofa-mosn/pkg/stream/http"
