+++ pkg/upstream/cluster/concreteclusters.go
@@ -20,7 +20,7 @@ package cluster
 import (
 	"net"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
 	"github.com/alipay/sofa-mosn/pkg/types"
 )
