+++ pkg/upstream/cluster/resourcemanager.go
@@ -20,7 +20,7 @@ package cluster
 import (
 	"sync/atomic"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/types"
 )
