+++ pkg/upstream/healthcheck/factory.go
@@ -18,7 +18,7 @@
 package healthcheck
 
 import (
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/protocol"
 	"github.com/alipay/sofa-mosn/pkg/types"
 )
