+++ pkg/upstream/healthcheck/sofarpchealthchecker.go
@@ -22,7 +22,7 @@ import (
 	"reflect"
 	"strconv"
 
-	"github.com/alipay/sofa-mosn/internal/api/v2"
+	"github.com/alipay/sofa-mosn/pkg/api/v2"
 	"github.com/alipay/sofa-mosn/pkg/log"
 	"github.com/alipay/sofa-mosn/pkg/protocol"
 	"github.com/alipay/sofa-mosn/pkg/protocol/sofarpc"
